+++ lib/models/pluginDependency.js
@@ -95,8 +95,8 @@ PluginDependency.listFromArray = function(arr) {
                 return PluginDependency.createFromString(entry);
             } else {
                 return PluginDependency({
-                    name: entry.name,
-                    version: entry.version
+                    name: entry.get('name'),
+                    version: entry.get('version')
                 });
             }
         })
